feat: Security Update: Vehicle Errors manage

+++ ACME.CargoExpress.API/Registration/Domain/Model/Commands/UpdateVehicleCommand.cs
@@ -1,3 +1,3 @@
 ﻿namespace ACME.CargoExpress.API.Registration.Domain.Model.Commands;
 
-public record UpdateVehicleCommand(int VehicleId, string Name, string Model, string Plate, string TractorPlate, float MaxLoad, float Volume, int EntrepreneurId);
+public record UpdateVehicleCommand(int VehicleId, string Name, string Model, string Plate, string TractorPlate, decimal MaxLoad, decimal Volume, int EntrepreneurId);
