+++ ACME.CargoExpress.API/Registration/Interfaces/REST/Resources/CreateVehicleResource.cs
@@ -1,3 +1,3 @@
 ﻿namespace ACME.CargoExpress.API.Registration.Interfaces.REST.Resources;
 
-public record CreateVehicleResource(string Name, string Model, string Plate, string TractorPlate, float MaxLoad, float Volume, int EntrepreneurId);
+public record CreateVehicleResource(string Name, string Model, string Plate, string TractorPlate, decimal MaxLoad, decimal Volume, int EntrepreneurId);
